Fix pre-existing problem with handling of end of windows
Due to a bug with the old version of the window package that audiobait
was depending on and a bug in audiobait, audiobait would spin in a
tight loop and emit a bunch of logs for the last minute of each day's
schedule.

The window package has now been upgraded and the audiobait's daily
playback logic has been fixed to stop trying to play combos which are
already done.